Safari-609.1.11.1/LayoutTests/http/tests/resourceLoadStatistics/downgraded-referrer-for-navigation-with-link-query-from-prevalent-resource-expected.txt

Check that document.referrer is downgraded to the eTLD+1 after a cross-site navigation with link query.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The referrer is downgraded.
PASS successfullyParsed is true

TEST COMPLETE
